Create midterm2.pde
1. Read about minim library
2. learned about "packages" that analyze sound files in different ways
3. understood visualization of sound (left and right speaker)
4. Still need to read more about Frequencies
5.  Drew some lines and shapes
6. played with Beatdetect.

I need to figure out how to play with colors and shapes.  how to analyze the audio file into frequency (or parts of the song) and create a visual language that symbolizes the different components of the song.  maybe it's a landscape? probably it will be something abstract.  I want to take some of the designs I played around with in HW 1 and HW 3.