Create midterm3.pde
I spoke to Che Yu about the possibility of setting up a variable that relates to the player.position.  We set up the arcs in the circle to rorate with player.position but I need to experiment with it more to understand how it works.

we also talked about playing around with Audio Analysis from Sound library to create variables that correlate to the different frequencies

here Che Yu showed me how to set up the buffersize variable to get a spiral effect, where the canvas is rotating at each point.  but i don't quite understand