Create midterm_broken2.pde
I was trying to do too many things, and then broke the detector function that I wrote in originally.  I was trying to create a a circle whose color changed with the beat.  I was also trying to create objects that did something at specific times in the song as defined by posPercentage